parse: don't reverse the order of resource records in the parser
Previously if we parsed a list of resource records which looked like this
in the packet:

  ;; ANSWER SECTION:
  download.camlcity.org.	299	IN	CNAME	cx30.rm.camlcity.org.
  cx30.rm.camlcity.org.	299	IN	A	138.201.190.144

We would get a list which `marshal` would convert into this:

  ;; ANSWER SECTION:
  cx30.rm.camlcity.org.	299	IN	A	138.201.190.144
  download.camlcity.org.	299	IN	CNAME	cx30.rm.camlcity.org.

Unfortunately some resolvers will reject this response. This was noticed
by downloading findlib-1.6.2 in an Ubuntu-12.04 container using an
ocaml-dns-derived DNS forwarder.

In general I think we should aim to make `parse ( marshal x ) = x`

This patch adjusts the test cases to match, since they were doing `List.rev`
themselves after calling `parse`.

Signed-off-by: David Scott <dave@recoil.org>